+++ packages/categorize/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.8.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@9.8.1...@pie-element/categorize@9.8.2) (2025-01-22)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-4452, PD-3872, PD-4689, PD-4704, PD-4588, PD-4698, PD-4618 ([a97329c](https://github.com/pie-framework/pie-elements/commit/a97329cf73619a3a5aae8d492edb8d6e5c5730df))
+
+
+
+
+
+## [9.8.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@9.8.0...@pie-element/categorize@9.8.1) (2025-01-22)
+
+
+### Bug Fixes
+
+* build ([68d8029](https://github.com/pie-framework/pie-elements/commit/68d8029520e531ae0a1eae15da6371289cde4c7a))
+* replace mra package with mr, includes accessibility PD-4527 ([c83bd43](https://github.com/pie-framework/pie-elements/commit/c83bd438d3a0e94d524c04364fddd3a8edc8876e))
+
+
+
+
+
 # [9.8.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@9.7.0...@pie-element/categorize@9.8.0) (2025-01-06)